requirements.txt
kafka-python<=2.0.2
confluent-kafka
six<=1.16.0  # Aggiunto esplicitamente per evitare conflitti con confluent-kafka

# Spark
pyspark

# Utility
pandas
numpy
matplotlib
requests
psutil

# Prometheus
prometheus-client